refactor(cache): split type tests for GH727
