+++ specs/GH727/tasks.md
@@ -174,24 +174,30 @@ GH-727 / #727
 - [x] `SP727-T161` Owner: coordinator. Done when: `src/core/providers/bedrock/provider_tests.rs` keeps only shared helpers and child module declarations. Verify: `sed -n '1,80p' src/core/providers/bedrock/provider_tests.rs`.
 - [x] `SP727-T162` Owner: coordinator. Done when: original Bedrock provider tests move into behavior-domain child modules without assertion changes. Verify: `rg -n "test_bedrock_provider_creation|test_messages_to_prompt_simple_user_message|test_transform_request_claude|test_transform_response_claude|test_calculate_cost_known_model" src/core/providers/bedrock/provider_tests.rs src/core/providers/bedrock/provider_tests/*.rs`.
 - [x] `SP727-T163` Owner: verification owner. Done when: all touched Bedrock provider test files are below U-16's 800-line ceiling and focused Bedrock provider tests pass. Verify: `wc -l src/core/providers/bedrock/provider_tests.rs src/core/providers/bedrock/provider_tests/*.rs`; `cargo test core::providers::bedrock::provider_tests --lib --all-features`.
-- [ ] `SP727-T164` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the Bedrock provider tests tranche. Verify: `cargo fmt --all -- --check`; `git diff --check`; `python3 /Users/apple/Desktop/code/AI/tool/specrail/checks/check_workflow.py --repo /Users/apple/Desktop/code/AI/tool/specrail --spec-dir "$PWD/specs/GH727"`; `cargo check --lib --all-features`; `cargo check --all-features --locked`; `cargo check`; GitHub PR CI and review-thread query.
+- [x] `SP727-T164` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the Bedrock provider tests tranche. Verify: #852 PR body, green PR CI, GraphQL reviewThreads totalCount 0, merge commit `4c278e8ab25e`, and #727 update comment.
+- [x] `SP727-T165` Owner: coordinator. Done when: after the Bedrock provider tests tranche merges, the next #727 tranche is selected from the remaining queue with fresh tracked-file line-count evidence. Verify: at `origin/main@4c278e8ab25e`, `src/core/cache/types.rs` is 861 lines and 24 tracked Rust files remain over the U-16 ceiling.
+- [x] `SP727-T166` Owner: coordinator. Done when: the cache types tranche documents why this slice is a test extraction rather than a production type facade split. Verify: `git diff -- specs/GH727/product.md specs/GH727/tech.md`.
+- [x] `SP727-T167` Owner: coordinator. Done when: `src/core/cache/types.rs` keeps production cache type definitions and delegates tests with `#[path = "types_tests.rs"] mod tests;`. Verify: `tail -n 20 src/core/cache/types.rs`.
+- [x] `SP727-T168` Owner: coordinator. Done when: original inline cache type tests move to `src/core/cache/types_tests.rs` without assertion changes. Verify: `rg -n "test_cache_key_new|test_dual_cache_config_default|test_atomic_cache_stats_concurrent|test_cache_stats_snapshot_combined" src/core/cache/types_tests.rs`.
+- [x] `SP727-T169` Owner: verification owner. Done when: both touched cache type files are below U-16's 800-line ceiling and focused cache type tests pass. Verify: `wc -l src/core/cache/types.rs src/core/cache/types_tests.rs`; `cargo test core::cache::types --lib --all-features`.
+- [ ] `SP727-T170` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the cache types tranche. Verify: `cargo fmt --all -- --check`; `git diff --check`; `python3 /Users/apple/Desktop/code/AI/tool/specrail/checks/check_workflow.py --repo /Users/apple/Desktop/code/AI/tool/specrail --spec-dir "$PWD/specs/GH727"`; `cargo check --lib --all-features`; `cargo check --all-features --locked`; `cargo check`; GitHub PR CI and review-thread query.
 
 ## 并行拆分
 
-This is a serial writable lane for the Bedrock provider test file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
+This is a serial writable lane for the cache types file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
 
 Writable ownership for this lane:
 
 - `specs/GH727/`
-- `src/core/providers/bedrock/provider_tests.rs`
-- `src/core/providers/bedrock/provider_tests/*.rs`
+- `src/core/cache/types.rs`
+- `src/core/cache/types_tests.rs`
 
 ## 验证
 
 - SpecRail packet review.
 - `cargo fmt --all -- --check`
 - `git diff --check`
-- `cargo test core::providers::bedrock::provider_tests --lib --all-features`
+- `cargo test core::cache::types --lib --all-features`
 - `cargo check --lib --all-features`
 - `cargo check --all-features --locked`
 - `cargo check`
@@ -199,5 +205,5 @@ Writable ownership for this lane:
 
 ## Handoff Notes
 
-This PR is the next #727 Bedrock provider test-suite split tranche and should use `Refs #727`, not `Closes #727`.
+This PR is the next #727 cache types test-extraction tranche and should use `Refs #727`, not `Closes #727`.
 The issue should remain open until the final scan shows no Rust files over the U-16 ceiling.
